Refactor EventHandler to EventDispatcher
