+++ src/main/java/ru/sbt/mipt/oop/event/EventGenerator.java
@@ -1,8 +1,5 @@
 package ru.sbt.mipt.oop.event;
 
-import ru.sbt.mipt.oop.event.SensorEvent;
-import ru.sbt.mipt.oop.event.SensorEventType;
-
 public interface EventGenerator {
 
     public static SensorEvent getNextSensorEvent() {
